requirements/base.txt
bleach==2.0.0
Django==1.11.6
django-autocomplete-light==3.2.10
django-contrib-comments==1.8.0
django-ditto==0.6.2
django-markdownx==2.0.21
django-spectator==3.3.0
django-taggit==0.22.0
psycopg2==2.7.1
pylibmc==1.5.2
pytz==2017.2
smartypants==2.0.0
whitenoise==4.0b4
